Change to have a specific value input for each account
In order to support the additional of FX rates the accounts need to be
tracked in their own home currency. This means that transferring $10
isn't the same to one account as the other. So, we can apply the FX rate
in the UI and just write the two different amounts to respective
accounts on the ledger.